requirement.txt
requests==2.28.2
beautifulsoup4==4.11.1
google-generativeai==0.8.5
python-dotenv==0.21.0
google-search-results